[sparkle] - refactor: integrate AnchoredPopover into Popover component
 - Removed the standalone AnchoredPopover component file and integrated its functionality into the Popover component file
 - Exported the AnchoredPopover directly from the Popover file to streamline the component structure
 - Updated AnchoredPopover imports in storybook to reflect the new location within the Popover component file
