+++ sparkle/src/components/Popover.tsx
@@ -9,7 +9,7 @@ const PopoverTrigger = PopoverPrimitive.Trigger;
 const PopoverPortal = PopoverPrimitive.Portal;
 const PopoverAnchor = PopoverPrimitive.Anchor;
 
-interface PopoverContentProps
+export interface PopoverContentProps
   extends React.ComponentPropsWithoutRef<typeof PopoverPrimitive.Content> {
   fullWidth?: boolean;
   mountPortal?: boolean;
@@ -100,6 +100,88 @@ function Popover({
 }
 
 PopoverContent.displayName = PopoverPrimitive.Content.displayName;
+
+interface AnchoredPopoverProps {
+  open: boolean;
+  anchorRef?: React.RefObject<HTMLElement>;
+  children: React.ReactNode;
+  popoverProps: PopoverContentProps;
+}
+
+export function AnchoredPopover({
+  open,
+  anchorRef,
+  children,
+  popoverProps,
+}: AnchoredPopoverProps) {
+  const [position, setPosition] = useState({
+    top: "50%",
+    left: "50%",
+    width: "0px",
+    height: "0px",
+  });
+
+  useEffect(() => {
+    if (!open) {
+      return;
+    }
+
+    const updatePosition = () => {
+      if (!anchorRef?.current) {
+        setPosition({
+          top: "50%",
+          left: "50%",
+          width: "0px",
+          height: "0px",
+        });
+        return;
+      }
+
+      const rect = anchorRef.current.getBoundingClientRect();
+      setPosition({
+        top: `${rect.top}px`,
+        left: `${rect.left}px`,
+        width: `${rect.width}px`,
+        height: `${rect.height}px`,
+      });
+    };
+
+    updatePosition();
+
+    const resizeObserver = new ResizeObserver(updatePosition);
+    if (anchorRef?.current) {
+      resizeObserver.observe(anchorRef.current);
+    }
+
+    window.addEventListener("scroll", updatePosition, true);
+
+    return () => {
+      resizeObserver.disconnect();
+      window.removeEventListener("scroll", updatePosition, true);
+    };
+  }, [open, anchorRef]);
+
+  return (
+    <PopoverRoot open={open} modal={false}>
+      <PopoverAnchor
+        className="s-fixed s-transition-all s-duration-300 s-ease-in-out"
+        style={{
+          top: position.top,
+          left: position.left,
+          width: position.width,
+          height: position.height,
+        }}
+      />
+      <PopoverContent
+        {...popoverProps}
+        onOpenAutoFocus={(e) => e.preventDefault()}
+        mountPortal={false}
+      >
+        {children}
+      </PopoverContent>
+    </PopoverRoot>
+  );
+}
 
 export {
   Popover,
